pyproject.toml
[project]
name = "mcp-tmdb"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
    "google-genai>=1.64.0",
    "langgraph>=1.0.9",
    "pytest>=9.0.2",
]
